CHE-6041 Move integration tests to che
