+++ assembly/assembly-wsmaster-war/src/main/resources/META-INF/persistence.xml
@@ -22,8 +22,9 @@
         <class>org.eclipse.che.api.user.server.model.impl.ProfileImpl</class>
         <class>org.eclipse.che.api.user.server.jpa.PreferenceEntity</class>
 
-        <class>com.codenvy.organization.spi.impl.OrganizationImpl</class>
-        <class>com.codenvy.organization.spi.impl.MemberImpl</class>
+        <class>org.eclipse.che.multiuser.organization.spi.impl.OrganizationImpl</class>
+        <class>org.eclipse.che.multiuser.organization.spi.impl.MemberImpl</class>
+
         <class>org.eclipse.che.api.workspace.server.model.impl.WorkspaceImpl</class>
         <class>org.eclipse.che.api.workspace.server.model.impl.WorkspaceConfigImpl</class>
         <class>org.eclipse.che.api.workspace.server.model.impl.ProjectConfigImpl</class>
@@ -61,7 +62,7 @@
 
         <class>org.eclipse.che.multiuser.resource.spi.impl.ResourceImpl</class>
         <class>org.eclipse.che.multiuser.resource.spi.impl.FreeResourcesLimitImpl</class>
-        <class>com.codenvy.organization.spi.impl.OrganizationDistributedResourcesImpl</class>
+        <class>org.eclipse.che.multiuser.organization.spi.impl.OrganizationDistributedResourcesImpl</class>
         <class>com.codenvy.api.license.server.model.impl.SystemLicenseActionImpl</class>
         <class>com.codenvy.api.invite.InviteImpl</class>
         <exclude-unlisted-classes>true</exclude-unlisted-classes>
