--- wsmaster/codenvy-hosted-integration-tests/codenvy-integration-cascade-jpa/src/test/resources/META-INF/persistence.xml
@@ -1,77 +1,0 @@
-<!--
-
-    Copyright (c) [2012] - [2017] Red Hat, Inc.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
-    Contributors:
-      Red Hat, Inc. - initial API and implementation
-
--->
-<persistence xmlns="http://java.sun.com/xml/ns/persistence"
-             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://java.sun.com/xml/ns/persistence persistence_1_0.xsd" version="1.0">
-    <persistence-unit name="main" transaction-type="RESOURCE_LOCAL">
-        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-        <class>org.eclipse.che.account.spi.AccountImpl</class>
-        <class>org.eclipse.che.api.user.server.model.impl.UserImpl</class>
-        <class>org.eclipse.che.api.user.server.model.impl.ProfileImpl</class>
-        <class>org.eclipse.che.api.user.server.jpa.PreferenceEntity</class>
-
-        <class>org.eclipse.che.api.workspace.server.model.impl.WorkspaceImpl</class>
-        <class>org.eclipse.che.api.workspace.server.model.impl.WorkspaceConfigImpl</class>
-        <class>org.eclipse.che.api.workspace.server.model.impl.ProjectConfigImpl</class>
-        <class>org.eclipse.che.api.workspace.server.model.impl.EnvironmentImpl</class>
-        <class>org.eclipse.che.api.workspace.server.model.impl.EnvironmentRecipeImpl</class>
-        <class>org.eclipse.che.api.workspace.server.model.impl.ExtendedMachineImpl</class>
-        <class>org.eclipse.che.api.workspace.server.model.impl.ProjectConfigImpl$Attribute</class>
-        <class>org.eclipse.che.api.workspace.server.model.impl.SourceStorageImpl</class>
-        <class>org.eclipse.che.api.workspace.server.model.impl.ServerConf2Impl</class>
-        <class>org.eclipse.che.api.workspace.server.model.impl.stack.StackImpl</class>
-
-        <class>org.eclipse.che.api.machine.server.model.impl.MachineSourceImpl</class>
-        <class>org.eclipse.che.api.machine.server.model.impl.CommandImpl</class>
-        <class>org.eclipse.che.api.machine.server.model.impl.SnapshotImpl</class>
-        <class>org.eclipse.che.api.machine.server.recipe.RecipeImpl</class>
-
-        <class>org.eclipse.che.api.factory.server.model.impl.FactoryImpl</class>
-        <class>org.eclipse.che.api.factory.server.model.impl.OnAppClosedImpl</class>
-        <class>org.eclipse.che.api.factory.server.model.impl.OnProjectsLoadedImpl</class>
-        <class>org.eclipse.che.api.factory.server.model.impl.OnAppLoadedImpl</class>
-        <class>org.eclipse.che.api.factory.server.model.impl.PoliciesImpl</class>
-        <class>org.eclipse.che.api.factory.server.model.impl.ActionImpl</class>
-        <class>org.eclipse.che.api.factory.server.model.impl.AuthorImpl</class>
-        <class>org.eclipse.che.api.factory.server.model.impl.ButtonAttributesImpl</class>
-        <class>org.eclipse.che.api.factory.server.model.impl.ButtonImpl</class>
-        <class>org.eclipse.che.api.factory.server.model.impl.IdeImpl</class>
-        <class>org.eclipse.che.api.factory.server.FactoryImage</class>
-
-        <class>org.eclipse.che.multiuser.api.permission.server.model.impl.SystemPermissionsImpl</class>
-        <class>org.eclipse.che.multiuser.api.permission.server.model.impl.AbstractPermissions</class>
-        <class>org.eclipse.che.multiuser.permission.workspace.server.model.impl.WorkerImpl</class>
-        <class>org.eclipse.che.multiuser.permission.workspace.server.stack.StackPermissionsImpl</class>
-        <class>org.eclipse.che.multiuser.permission.machine.recipe.RecipePermissionsImpl</class>
-        <class>org.eclipse.che.api.ssh.server.model.impl.SshPairImpl</class>
-
-        <class>com.codenvy.organization.spi.impl.OrganizationImpl</class>
-        <class>com.codenvy.organization.spi.impl.MemberImpl</class>
-        <class>com.codenvy.organization.spi.impl.OrganizationDistributedResourcesImpl</class>
-
-        <class>org.eclipse.che.multiuser.resource.spi.impl.ResourceImpl</class>
-        <class>org.eclipse.che.multiuser.resource.spi.impl.FreeResourcesLimitImpl</class>
-        <exclude-unlisted-classes>true</exclude-unlisted-classes>
-        <properties>
-            <property name="javax.persistence.jdbc.driver" value="org.h2.Driver"/>
-            <property name="javax.persistence.jdbc.url" value="jdbc:h2:mem:test"/>
-            <property name="javax.persistence.jdbc.user" value=""/>
-            <property name="javax.persistence.jdbc.password" value=""/>
-
-            <property name="eclipselink.exception-handler" value="org.eclipse.che.core.db.h2.jpa.eclipselink.H2ExceptionHandler"/>
-            <property name="eclipselink.target-server" value="None"/>
-            <property name="eclipselink.logging.logger" value="DefaultLogger"/>
-            <property name="eclipselink.logging.level" value="SEVERE"/>
-        </properties>
-    </persistence-unit>
-</persistence>
