--- wsmaster/codenvy-hosted-integration-tests/codenvy-integration-cascade-jpa/src/test/resources/logback-test.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) [2012] - [2017] Red Hat, Inc.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
-    Contributors:
-      Red Hat, Inc. - initial API and implementation
-
--->
-<configuration>
-    <appender name="stdout" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%-41(%date[%.15thread]) %-45([%-5level] [%.30logger{30} %L]) - %msg%n%nopex</pattern>
-        </encoder>
-    </appender>
-    <appender name="file" class="ch.qos.logback.core.FileAppender">
-        <File>target/log/test.log</File>
-        <encoder>
-            <pattern>%-41(%date[%.15thread]) %-45([%-5level] [%.30logger{30} %L]) - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-
-    <root level="ERROR">
-        <appender-ref ref="stdout"/>
-        <appender-ref ref="file"/>
-    </root>
-
-</configuration>
